Control lines
Add in code to handle most of the RS-232 control lines
The RI lines for USART1 and UART4 are not yet handled.
The code has not been tested and is not connected into
the usbcdc code.

Code also needs to be added in for when the control
lines recieve interrupts andand hardware flow control
is enabled (stop transmiting data and possibly stop
recieving data)

Signed-off-by: Joseph A. Lutz <CipherWizard@cipherwizard.net>