raise direction output while transmitting
Raise the direction output associated with each port when that port is
transmitting rather than raising it when RTS is set and lowering it when
RTS is cleared. This results in RS485 transmitter being turned on for
less time than if it's controlled with RTS which make it possible to
communicate with devices which reply in under 3ms.

Mantis ID: 5839

Signed-off-by: George McCollister <george.mccollister@gmail.com>