delay TX empty notificaton for 9600 baud rate as well
Still getting CRC errors that looked like the end of the message was
being cut off.
If some other interrupt is already using sys_tick_handler() to delay and
tx_empty_count_down is set to 1 on a uart, it is possible that the
systic will time out imediatly after it was set. This will not alow
adaquate time passing for the data to be compleatly shifted out of the
register before the hardware->dir.pin is changed. This will cut off any
more of the last byte from being transmitted.
To quickly get around the issue I changed the tx_empty_delay for 9600 to
2. This may make it take longer than expected to switch the
hardware->dir.pin from transmit to receive. This is just a quick fix for
an edge case.
A better solution would be to make systic interrupt more often and make
the interrupt loop more effecent, making it take less time.

Signed-off-by: Joseph Lutz <joseph.lutz@novatechweb.com>