Audjust the RX timers for 115200 baud and higher
The code was set to time out the RX buffer after a certain time period.
That time period was the amount of time it takes to send three
characters. I changed that timeout to six characters time length. This
gets around the amount problem where one character is lost ocasionaly.

Signed-off-by: Joseph Lutz <joseph.lutz@novatechautomation.com>